Implement expanding open-world board

+++ app/services/player-service.js
@@ -1,5 +1,4 @@
 'use strict';
-const Board = require('../configs/board');
 const ServerConfig = require('../configs/server-config');
 const Player = require('../models/player');
 
@@ -37,7 +36,7 @@ class PlayerService {
         const playerName = this.nameService.getPlayerName();
         const newPlayer = this.createPlayer(socket.id, playerName);
         socket.emit(ServerConfig.IO.OUTGOING.NEW_PLAYER_INFO, playerName, newPlayer.color);
-        socket.emit(ServerConfig.IO.OUTGOING.BOARD_INFO, Board);
+        socket.emit(ServerConfig.IO.OUTGOING.BOARD_INFO, this.boardOccupancyService.getBoardInfo());
         this.notificationService.broadcastNotification(`${playerName} entrou no jogo!`, newPlayer.color);
         const backgroundImage = this.imageService.getBackgroundImage();
         if (backgroundImage) {
@@ -165,10 +164,11 @@ class PlayerService {
             if (this.playerContainer.isSpectating(player.id)) {
                 continue;
             }
+            const nextCoordinate = CoordinateService.getNextCoordinate(player.getHeadCoordinate(), player.direction);
+            this.boardOccupancyService.ensureCoordinateWithinBounds(nextCoordinate);
             this.boardOccupancyService.removePlayerOccupancy(player.id, player.getSegments());
-            CoordinateService.movePlayer(player);
-            if (this.boardOccupancyService.isOutOfBounds(player.getHeadCoordinate()) ||
-                    this.boardOccupancyService.isWall(player.getHeadCoordinate())) {
+            player.move(nextCoordinate);
+            if (this.boardOccupancyService.isWall(player.getHeadCoordinate())) {
                 player.clearAllSegments();
                 this.playerContainer.addPlayerIdToRespawn(player.id);
                 this.notificationService.broadcastRanIntoWall(player.name, player.color);
